refactor: 기존 component 폴더를 pages로 변경
